requirements.txt
python-dotenv
streamlit
pandas
numpy